fix/refactor: resolve issues #901, #902, #903, #904
## #903 — Add OTEL_DEBUG production guard
In backend/src/tracing.ts, added a startup check that detects when
NODE_ENV=production and OTEL_DEBUG=true are set simultaneously. The
check logs a CRITICAL error explaining the performance risk and
programmatically disables debug mode, preventing per-span I/O latency
logging from running in production.

## #902 — Fix ExportService Content-Length for streaming responses
Removed the pre-calculated Content-Length header from all four streaming
export methods in backend/src/services/ExportService.ts (analytics CSV,
analytics NDJSON, posts CSV, posts NDJSON). Set Transfer-Encoding:
chunked instead via the shared makeStream helper. This prevents HTTP
clients from truncating responses when actual row sizes exceed the
estimate.

## #901 — Audit and standardize permission checks across all routes
Audited all route files in backend/src/routes/ and applied
authMiddleware + checkPermission to every state-changing endpoint that
was missing explicit authorization. Changes per route:
- ai.ts: added checkPermission('posts:create') alongside existing auth
- analytics.ts: added authMiddleware + checkPermission('analytics:view')
- circuitBreaker.ts: added auth + checkPermission('settings:manage') to
  POST /reset, POST /:service/open, POST /:service/close
- exports.ts: added auth + checkPermission('analytics:export') to
  GET /analytics and GET /posts
- facebook.ts: added auth + checkPermission('posts:create') to POST
  /post and POST /comment/:id/reply; checkPermission('posts:delete') to
  DELETE /comment/:id
- images.ts: added auth + checkPermission('posts:create') to POST
  /upload; auth + checkPermission('settings:manage') to DELETE /cache
- jobs.ts: added auth + checkPermission('settings:manage') to all
  mutating endpoints (retry, retry-all, delete, pause, resume, clear)
- linkedin.ts: added auth + checkPermission('posts:create') to POST
  /share
- tiktok.ts: added auth + checkPermission('posts:create') to POST
  /video/upload and POST /video/upload-url
- translation.ts: added auth + checkPermission('posts:create') to POST
  /translate, POST /detect, POST /batch
- video.ts: added auth + checkPermission('posts:create') to POST
  /upload; auth + checkPermission('posts:delete') to DELETE /job/:jobId
Each route documents its required permission in a comment.

## #904 — Split IPFSService into focused modules
Refactored src/blockchain/services/IPFSService.ts (16 KB monolith) into
three independently testable modules:
- IPFSClient.ts: connection management, provider config, IndexedDB
  helpers, retry/timeout utilities
- IPFSUploader.ts: file upload (simple + chunked), JSON upload, batch
  upload, pin management, pinned file listing
- IPFSRetriever.ts: file retrieval, JSON retrieval, gateway fallback,
  cache management (store/evict/clear), unpin
IPFSService.ts is kept as a backward-compatible facade that composes the
three modules and delegates all calls, preserving the existing public API.

Closes #901
Closes #902
Closes #903
Closes #904
