+++ backend/src/routes/images.ts
@@ -1,6 +1,8 @@
 import { Router, Request, Response } from 'express';
 import multer from 'multer';
 import path from 'path';
+import { authenticate as authMiddleware } from '../middleware/authenticate';
+import { checkPermission } from '../middleware/checkPermission';
 import { ImageOptimizationService } from '../services/ImageOptimizationService';
 
 const router = Router();
@@ -78,7 +80,8 @@ const upload = multer({
  *       400:
  *         description: No image provided
  */
-router.post('/upload', upload.single('image'), async (req: Request, res: Response) => {
+// Required permission: posts:create
+router.post('/upload', authMiddleware, checkPermission('posts:create'), upload.single('image'), async (req: Request, res: Response) => {
   try {
     if (!req.file) {
       return res.status(400).json({ error: 'No image provided' });
@@ -223,7 +226,8 @@ router.get('/cache/size', async (req: Request, res: Response) => {
  *       200:
  *         description: Cache cleared
  */
-router.delete('/cache', async (req: Request, res: Response) => {
+// Required permission: settings:manage
+router.delete('/cache', authMiddleware, checkPermission('settings:manage'), async (req: Request, res: Response) => {
   try {
     await ImageOptimizationService.clearCache();
     res.json({ message: 'Cache cleared' });
